% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistBoxplot.R
\name{DistBoxplot}
\alias{DistBoxplot}
\title{DistBoxplot}
\usage{
DistBoxplot(
  dm,
  group,
  dm_name = "",
  group_name = "",
  IndividualID = NULL,
  outpath = "./"
)
}
\arguments{
\item{dm}{A squared distance matrix}

\item{group}{A metadata variable corresponds to samples in the distance matrix.}

\item{dm_name}{An optional character string for the distance name.}

\item{group_name}{An optional character string for the metadata variable name.}

\item{IndividualID}{Another metadata variable that can make this function return only paired distance values within this group.}

\item{outdir}{The output directory.}
}
\description{
DistBoxplot
}
\examples{
set.seed(123)
x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
x0 <- data.frame(rbind(t(rmultinom(7, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(8, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289)))))
y<-factor(c(rep("A", 30), rep("B", 30)))
y0<-factor(c(rep("A", 5), rep("B", 55)))
require(vegan)
dm<-vegdist(x, method="bray")
dm<-data.matrix(dm)
DistBoxplot(dm, y, dm_name="bray", group="y")
}
\author{
Shi Huang
}
