% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataChecker.R
\name{effect_size_eval}
\alias{effect_size_eval}
\title{effect_size_eval}
\usage{
effect_size_eval(x, metadata)
}
\arguments{
\item{x}{A dataframe containing the biolgical data}

\item{metadata}{A dataframe with > two columns corresponds to samples (rownames) in the biological data.}
}
\description{
effect_size_eval
}
\author{
Shi Huang
}
