% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistSplit.R
\name{split_distance_matrix_by_metadata}
\alias{split_distance_matrix_by_metadata}
\title{DistBoxplot}
\usage{
split_distance_matrix_by_metadata(dm, metadata, split_factor)
}
\arguments{
\item{dm}{A squared distance matrix}

\item{group}{A metadata variable corresponds to samples in the distance matrix.}

\item{dm_name}{An optional character string for the distance name.}

\item{group_name}{An optional character string for the metadata variable name.}

\item{IndividualID}{Another metadata variable that can make this function return only paired distance values within this group.}

\item{outdir}{The output directory.}
}
\description{
DistBoxplot
}
\examples{
set.seed(123)
x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
x0 <- data.frame(rbind(t(rmultinom(7, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(8, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289)))))
y<-factor(c(rep("A", 30), rep("B", 30)))
z<-factor(c(rep("A", 20), rep("B", 20), rep("C", 20)))
metadata<-data.frame(y, z)
require(vegan)
dm<-vegdist(x, method="bray")
split_distance_matrix_by_metadata(dm, metadata, split_factor="y")
}
\author{
Shi Huang
}
